+++ connectlib/strategies/strategy.py
@@ -8,6 +8,7 @@ from connectlib.algorithms.algo import Algo
 from connectlib.nodes.aggregation_node import AggregationNode
 from connectlib.nodes.test_data_node import TestDataNode
 from connectlib.nodes.train_data_node import TrainDataNode
+from connectlib.schemas import StrategyName
 
 SharedState = TypeVar("SharedState")
 
@@ -18,6 +19,16 @@ class Strategy(ABC):
     def __init__(self, *args, **kwargs):
         self.args = args
         self.kwargs = kwargs
+
+    @property
+    @abstractmethod
+    def name(self) -> StrategyName:
+        """The name of the strategy
+
+        Returns:
+            StrategyName: Name of the strategy
+        """
+        raise NotImplementedError
 
     @abstractmethod
     def perform_round(